refactor(uploader): split uploader.go into focused files
uploader.go had grown past 1000 lines and mixed planning, transferring,
retrying, remote path handling, dialing and host key verification in one
place. Split it along those seams:

- planner.go: fileItem/plan, buildPlan, dirsForFiles, hashPlanFiles
- transfer.go: uploadFiles, uploadFile, renameReplace, cleanupTmp, ctxReader
- retry.go: uploadFileWithRetry, isRetryable, sleepCtx
- remote.go: remote dir creation, listing, path helpers, delete guards
- connection.go: hop, connect, dialViaJump, authMethods, keepalives
- hostkeys.go: hostKeyCallback, knownHostsCallback

uploader.go keeps Logger, Stats, Run and the per-strategy dispatch, and
its package comment now maps out which file holds what.

Pure code movement: no declaration was renamed, changed or removed, and
a sorted line-level diff of the old file against the new set shows only
the repeated per-file imports and the extended package comment as
additions.
